Merge pull request #5753 from RosettaCommons/vmullig/string_copy_issue
Fixing some strings that were passed by copy instead of by reference.

Strings are relatively heavy objects, and copying them just to pass them to a function is a waste of memory and clock cycles. These ones are in some utility functions. This is an easy issue for newer devs to review, and it's ready for review.

Old repository SHA1: 15591c560bdc0903f1c4731934a231d8279f4f0c