Merge pull request #6076 from RosettaCommons/vmullig/fix_symmetrical_residue_selector
Fix the SymmetricalResidueSelector's internal logic

The SymmetricalResidueSelector would copy selections from the independent set to the dependent residues, but not vice versa.  This PR fixes this, so that it's guaranteed to produce a symmetrical selection every time.  This requires a two-pass process:
- In the first pass, every independent residue that has a dependent selected is selected.
- In the second pass, every independent residue that is selected selects all of its dependents.

Old repository SHA1: f0932dcdc69a51e5c803c8e1926283655728c935