Merge pull request #5867 from RosettaCommons/roccomoretti/packer_task_thresholds
Make Dunbrack probability thresholds non-global.



Right now, the -packing::dunbrack_prob_buried options and the like are a set-once option in the dunbrack/RotamerLibrary singleton.

It looks like it doesn't need to be. Instead, as it controls the Rotamer generation logic (e.g. which rotamers are selected).
we can move these options to the PackerTask.

It's not 100% clean, as to preserve the current behavior we can't put the option reading under the same control as the other command line options, but at least it allows us to change these thresholds on a per-packing run basis, rather than a once-per-program-launch basis. (This will help with certain use cases in Foldit and PyRosetta.)

Old repository SHA1: 81b8018c4a6220108ae82919ff0011d098eeb590