Merge pull request #5360 from RosettaCommons/roccomoretti/chemistry_objects
Add "Chemistry" Objects, which are like Movers, but for ResidueTypes. This is part of the process of merging my the drug_design branch piecemeal.

The Chemisty objects come in two parts. The first is a simple core-level one, which defines the basic interface. The second is a protocols-level one which extents the interface to be XML parsable and Pose-aware. (That is, you can potentially have chemistries which are aware of the residue context in the Pose.) Chemistries occupy a new section in the input XML.

With this PR comes some simple Chemistries, such as PatchChemistry, which allows you to apply the same manipulations you could in a patch.

I also add an ApplyChemistyMover, which allows you to apply one or more Chemistries to a Pose from an XML context.

Old repository SHA1: d8d9982120b445e65d19d1e63ccc663cc1d639e8

Co-authored-by: brownbp1 <benjamin.p.brown17@gmail.com>