Merge pull request #6346 from RosettaCommons/roccomoretti/obey_ssbond
Add option to obey SSBOND info during PDB loading.

One of the time sinks in PDB loading is the disulfide bond detection. There is currently a way to turn it off entirely (-detect_disulf false) or use an external file to specify things (-fix_disulf), but there's not currently a way to specify that you want to obey the disulfide annotations in the PDB file and not do a general scan. This is less than ideal, as from-wwPDB and Rosetta output files should have the complete SSBOND information in them already.

This PR adds a -in:obey_ssbond option to the command line, which tells the loader code to take the SSBOND information from the PDB/StructFileRep as the only disulfides which should be used. Adding this flag shaves the time in initialize_disulfide_bonds() in half, saving somewhere between 1-2% in PDB loading speed. 

Old repository SHA1: 2490d4c2089049cce8054d497945c4b684a6c771