Adjust the git submodule paths for Forking potential.
We use relative paths for many of the RosettaCommons submodules, to
allow for https/ssh flexibility. However, for the open repository if
someone were working off a fork, they wouldn't be able to get the
submodules unless they also forked the submodule.

To fix that, we can be explicit about the RosettaCommons organization in
the submodule URL.