Documentation Updates for the Open Repository. (#2)
* Documentation Updates for the Open Repository.

When we go to the open repository, we need to add additional information
for public viewers, as well as update a fair amount of the current info
(e.g. README).

Here is the current version of those changes. More changes will likely
be coming as we think of more things to change.