Update the BCL to use the most recent version of the public repository.
In light of opening Rosetta (and especially in light of the BCL being
opened and switching repositories), it would be good to have Rosetta use
the public version of the BCL repository, rather than the old private
one.

Doing this is actually rather straightforward, effectively amounting to
re-running the rosetta_integration/update_bcl.py script (though with
some updates). There's some minor changes to the bcl.external.settings
file.

I've also added an integration test, to check if the integration works. (It does on my machine.)

The big caveat is where/how to host the updated BCL repository. The
commit with the associated SHA1 is in my fork of the public BCL repo at
the moment, but before merging we can change that.

PR RosettaCommons/main#6042 advocated doing some musical chairs with the old and new repo.
My thought is it's likely easiest if we just go with a new repo name --
that is, clone the BCLCommons/bcl repo as something like
RosettaCommons/bcl_public, and then use that. That preserves the use of
the old submodule SHA1s under the old repo name in history if anyone
needs to use them, but allows us to use the new repo for future commits.
The only downside is that the repo name on the BCLCommons and on
RosettaCommons won't be exactly the same. (But I think it will be close
enough not to matter.)

If such course of action is alright by people (particularly @brownbp1 &
@lyskov who were active on 6042), then I can do the fork and push the
relevant commit to the new repo name.