Clean up pilot apps (#14)
Especially now that we'll be exposing pilot apps to 'regular' users with
the open repo, we probably don't need to be compiling all the pilot apps
that we are. The general concept here is that if the app isn't tested with an
integration test, it probably isn't important enough to be generally
compiled.

If you're seeing this in the future (e.g. from a git blame) feel free to
uncomment any of these apps, but I'd HIGHLY recommend also adding an
integration test with a short sample run, to avoid this sort of issue in
the future.