Update the BCL to use the most recent version of the public repository. (#15)
In light of opening Rosetta (and especially in light of the BCL being
opened and switching repositories), it would be good to have Rosetta use
the public version of the BCL repository, rather than the old private
one.

Doing this is actually rather straightforward, effectively amounting to
re-running the rosetta_integration/update_bcl.py script (though with
some updates).  We are now using the RosettaCommons/bcl_rosetta repo 
to host the modified version of the BCL.

I've also added an integration test, to check if the integration works. (It does on my machine.)

There's still some lingering issues with Linux+Clang compilation, but since GCC seems to be working fine, we can handle those at a later date.