Fix issue with CartesianBondedEnergy.
Calling rsd1.upper_connect() is not well formed if the residue is not an upper connection.
CartesianBondedEnergy makes the assumption that if you're not an upper terminus, you have an upper connect.
This isn't always the case with more esoteric residues. But if we combine with the is-protein check that is being done anyway,
we can avoid the majority of the cases where this may be a problem.