Crosslinker Mover lanthionine (#24)
This is the code to enable lanthipeptide modeling in Rosetta. 
Crosslinker updated to include lanthionine and methyllanthionine connections.
-This includes D/L ALA steriochem for lanthionine and D/L and 3 R/S steriochem for methylanthionine (B(L/D)(R/S))
Dehydrated amino acid patches for DBU and DHU were updated/added
The AmbiguouosNMRDistanceConstraints were updated to work for NCAA (hardcoding was generalized)

Additionally, I noticed that the thioether lariat in simple cyc pep predict only samples cis omega bonds for the N terminal residue. I changed this to trans, which should be energetically preferred.