Add debug printing for inline residue selector parsing logic.
There was some discussion that the inline logic for ResidueSelectors doesn't always do the same thing as the separate AND/OR/NOT
selectors do. This PR adds a rudimentary nested residue selector printing facility, and enables it with debug output on the core.select.residue_selector.util tracer
when parsing inline logic.

It's primarily for debug purposes, and may or may not be useful in practice.