Monte Carlo Drug Design Movers.
This is code from the long-running drug design branch (PR 1036 in the old repo),
and consists of everything which hasn't already been merged and is probably worth still merging.

This is being merged at this time primarily as Tracy Tang based her work on this branch,
and is getting ready to merge & publish her stuff.