Bugfix for RosettaCM and branched glycans (#221)
RosettaCM, with a branched glycan ligand, was segfaulting. In 2 locations, the "previous residue" of the given polymer entity was selected by choosing the position prior to the one in question, which is not appropriate for a branched glycan. The fix was to set a conditional for carbohydrates that uses conformation::carbohydrates::find_seqpos_of_saccharides_parent_residue to find the previous residue.