Monte Carlo Drug Design, base code. (#282)
This is code from the long-running drug design branch (PR 1036 in the old repo), and consists of everything which hasn't already been merged and is probably worth still merging.

This is being merged at this time primarily as Tracy Tang has based her work on this branch, and is getting ready to merge & publish her stuff.

The changes are primarily Movers, Filters and and Chemistries added to src/protocols/drug_design/

Other changes of notes are additions of the src/protocols/rotamer_gen/ directory, which contains two Chemistry objects, one which adds RDKit rotamers, and one which adds BCL rotamers (and requires the extras=bcl build).