Updated Ligand Motifs Code (#59)
The code I am contributing here are updates to improve the functionality of the ligand motif code. This includes improvements to the following protocols:

motifs/LigandDiscoverySearch - Created a method for scoring placed ligands based on how much of a user-defined binding pocket is filled, improved use of score functions, collection and analysis of motifs collected off of placed ligands to confirm contacts against residues of interest

motifs/IdentifyLigandMotifs - Moved usage of process_for_motifs from private to public so that the function to collect motifs can be used on poses in runtime (and not exclusively from imported files). The process_for_motifs function is now used in LigandDiscoverySearch.

Added new options in options_rosetta.py under motifs for use in LigandDiscoverySearch.

Added public app remove_duplicate_motifs.cc, which removed motifs that are considered duplicates based on similar distance and angle of atoms involved. This helps improve runtimes of protocols that use lists of motifs by not considering the use of motifs that are effectively very similar/same. I have added an integration test as well.