Lanthipeptide and recon multistate design update (#263)
Modifications to lanthionine ring parameters to more accurately model the sidechain torsions and the methyl placement for methyllanthionine. Additionally, the recon multistate design code is updated to allow for design with NCAA by switching from a single letter code sequence stored in a string to a vector1 of strings that store the base name of the residue.