FakeRotLib and minor tweaks to m2pp.py (#349)
This pull request is to introduce the FakeRotLib script formally into the rosetta codebase. The script takes in a NCAA sdf, generates conformers using RDKit, parameterizes it using molfile_to_params_polymer.py, and optionally creates a rotlib file describing the rotamer distributions of the chi torsions.

A few small tweaks were also made to molfile_to_params_polymer.py to ensure that the backbone hydrogen is always described using a proper torsion and not an improper torsion and to ensure guanadinium nitrogens are properly typed.