Remove debug stdout (#482)
This quick PR deletes a statement printed to stdout when using the ProtocolSettingsMetric CompositeStringMetric.

With the `get_user_options="1"` option, the metric prints `Skip user options 1` to stdout, which is the opposite of what runs. Then, with the `get_user_options="0"` option, it prints `Skip user options 0`, which is again the opposite of what runs.