Fix dropped settings issues in HighResDocker
The copy constructor of HighResDocker was not copying over the resfile_ member, which means it was ignoring that setting.
Since the copy constructor is effectively a straight member-by-member copy, we can simply delete it and rely on the autogenerated copy constructor.

Additionally, I noticed that the initialize_from_options() function was declaring local variables, rather than changing the member variables. Fix this.